[MODFIN-166] - Refactor for using RestClient
